+++ src/chip/superfx/superfx.h
@@ -6,8 +6,8 @@ class SuperFX : public MMIO { public:
   void power();
   void reset();
 
-  uint8 mmio_read (uint16 addr);
-  void  mmio_write(uint16 addr, uint8 data);
+  uint8 mmio_read (uint addr);
+  void  mmio_write(uint addr, uint8 data);
 
 private:
   #include "regs.h"
