+++ src/cpu/scpu/mmio/mmio.h
@@ -1,7 +1,7 @@
   void   mmio_power();
   void   mmio_reset();
-  uint8  mmio_read (uint16 addr);
-  void   mmio_write(uint16 addr, uint8 data);
+  uint8  mmio_read (uint addr);
+  void   mmio_write(uint addr, uint8 data);
 
   uint8  pio_status();
