+++ src/ppu/bppu/bppu.h
@@ -220,8 +220,8 @@ struct {
   uint8  mmio_r213e();            //STAT77
   uint8  mmio_r213f();            //STAT78
 
-  uint8  mmio_read (uint16 addr);
-  void   mmio_write(uint16 addr, uint8 data);
+  uint8  mmio_read (uint addr);
+  void   mmio_write(uint addr, uint8 data);
 
   void   latch_counters();
